requirements.txt.txt
groq
python-telegram-bot==20.7
requests
pypdf
